# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.1.0"
  title: DataSourceAPI
# during dev, should point to your local machine
host: localhost:8080
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  # - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /datasources:
    get:
      description: Return all `DataSource`
      parameters:
        - $ref: "#/parameters/sortParam"
        - $ref: "#/parameters/skipParam"
        - $ref: "#/parameters/limitParam"
        - $ref: "#/parameters/queryParam"
        - $ref: "#/parameters/selectParam"
        - $ref: "#/parameters/distinctParam"
      responses:
        200: 
          description: Success
          schema: 
            $ref: "#/definitions/DataSource"
    post:
      description: Create new DataSource and configures it.
      parameters:
        - name: DataSource
          in: body
          required: true
          schema:
            $ref: "#/definitions/DataSource"
      responses:
        201:
          description: Success
          schema:
            $ref: "#/definitions/DataSource"
...
  /datasources/{_id}:
    get:
      description: Get one entity
      parameters:
        - $ref: "#/parameters/idParam"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/DataSource"
    delete:
      parameters:
        - $ref: "#/parameters/idParam"
      description: Delete one entity.
      responses:
        200:
          description: OK
    patch:
      description: Update an entity
      parameters:
        - name: DataSource
          in: body
          required: true
          schema:
            $ref: "#/definitions/DataSource"
        - $ref: "#/parameters/idParam"
      responses:
        200:
          description: OK
    # POST identical to PUT
    # post:
    #   description: Update an entity
    #   parameters:
    #     - name: DataSource
    #       in: body
    #       required: true
    #       schema:
    #         $ref: "#/definitions/DataSource"
    #     - $ref: "#/parameters/idParam"
    #   responses:
    #     200:
    #       description: OK
...
definitions:
...
  DataSource:
    required:
      - name
      - adapter
    properties:
      name:
        type: string
      adapter:
        type: string
      adapterProperties:
        type: object
...
parameters:
  idParam:
    name: _id
    in: path
    required: true
    type: string
  sortParam:
    name: sort
    in: query
    description: | 
      Sort entities by an attribute.
      Example: sort=name, sort=-name, sort={"name":-1}
    required: false
    type: string
  skipParam:
    name: skip
    in: query
    description: | 
      Number of entities to be skipped, like an offset.
      Example: skip=10
    required: false
    type: integer
  limitParam:
    name: limit
    in: query
    description: | 
      Limit number of entities returned
      Example: limit=10
    required: false
    type: integer
  queryParam:
    name: query
    in: query
    description: | 
      A MongoDB query.
      Example: query={"age":"<=12"}
    required: false
    type: string
  selectParam:
    name: select
    in: query
    description: | 
      Select attributes to be show or hidden
      Example: select=name,createdAt
    required: false
    type: string
  distinctParam:
    name: distinct
    in: query
    description: | 
      Attribute to be distinct
      Example: distinct=name
    required: false
    type: string
...